:sparkles: mv pkg/{runtime,manager}/signals
pkg/runtime is a bit of a odd collection right now.  This starts to move
stuff out of it, leaving aliases in place instead.  Since signals are
used to control managers, this functionality now lives in a subpackage
of manager.
