+++ pkg/manager/signals/doc.go
