+++ pkg/manager/signals/signal_posix.go
