+++ pkg/manager/signals/signal_test.go
