+++ pkg/manager/signals/signal_windows.go
