+++ pkg/manager/signals/signals_suite_test.go
